# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TariffZone = typeof TariffZone[keyof typeof TariffZone];


export const TariffZone = {
  zone1: 'zone1',
  zone2: 'zone2',
  zone3: 'zone3',
} as const;

export type RouteOption = typeof RouteOption[keyof typeof RouteOption];


export const RouteOption = {
  zone1: 'zone1',
  zone2: 'zone2',
  zone3: 'zone3',
  none: 'none',
} as const;

export type BookingType = typeof BookingType[keyof typeof BookingType];


export const BookingType = {
  full_year: 'full_year',
  first_half: 'first_half',
} as const;

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  received: 'received',
  reviewed: 'reviewed',
  confirmed: 'confirmed',
  query_open: 'query_open',
} as const;

export interface SiblingInput {
  childName: string;
  /** @nullable */
  studentNumber?: string | null;
  gradeYear: string;
  outboundRoute: RouteOption;
  returnRoute: RouteOption;
}

export interface BookingInput {
  /** @minLength 2 */
  childName: string;
  /** @minLength 5 */
  childAddress: string;
  /** @nullable */
  studentNumber?: string | null;
  gradeYear: string;
  /** @minLength 2 */
  parentName: string;
  parentEmail: string;
  /** @minLength 6 */
  parentPhone: string;
  tariffZone: TariffZone;
  bookingType: BookingType;
  outboundRoute: RouteOption;
  returnRoute: RouteOption;
  /** @maxItems 3 */
  siblings: SiblingInput[];
  confirmationAccepted: boolean;
  /** @minLength 2 */
  signatureName: string;
  gdprConsent?: boolean;
}

export interface BookingConfirmation {
  id: number;
  referenceNumber: string;
  message: string;
}

export interface Sibling {
  id: number;
  childName: string;
  /** @nullable */
  studentNumber?: string | null;
  gradeYear: string;
  outboundRoute: string;
  returnRoute: string;
}

export interface AdminBooking {
  id: number;
  referenceNumber: string;
  childName: string;
  gradeYear: string;
  parentName: string;
  parentEmail: string;
  parentPhone?: string;
  tariffZone: string;
  bookingType: string;
  outboundRoute: string;
  returnRoute: string;
  status: string;
  createdAt: string;
  siblingCount: number;
}

export interface AdminBookingDetail {
  id: number;
  referenceNumber: string;
  childName: string;
  childAddress: string;
  /** @nullable */
  studentNumber?: string | null;
  gradeYear: string;
  parentName: string;
  parentEmail: string;
  parentPhone: string;
  tariffZone: string;
  bookingType: string;
  outboundRoute: string;
  returnRoute: string;
  signatureName?: string;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
  siblings: Sibling[];
}

export interface AdminBookingList {
  bookings: AdminBooking[];
  total: number;
  page: number;
  limit: number;
}

export interface BookingStatusUpdate {
  status?: BookingStatus;
  /** @nullable */
  adminNotes?: string | null;
}

export type AdminStatsByStatus = {[key: string]: number};

export type AdminStatsByTariffZone = {[key: string]: number};

export type AdminStatsByBookingType = {[key: string]: number};

export type AdminStatsByGradeYear = {[key: string]: number};

export interface AdminStats {
  totalBookings: number;
  byStatus: AdminStatsByStatus;
  byTariffZone: AdminStatsByTariffZone;
  byBookingType: AdminStatsByBookingType;
  byGradeYear: AdminStatsByGradeYear;
  totalChildren: number;
  recentBookings: AdminBooking[];
}

export interface AdminCredentials {
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  username?: string | null;
}

export type ListAdminBookingsParams = {
/**
 * @nullable
 */
tariffZone?: string | null;
/**
 * @nullable
 */
bookingType?: string | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
gradeYear?: string | null;
/**
 * @nullable
 */
outboundRoute?: string | null;
/**
 * @nullable
 */
returnRoute?: string | null;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ExportBookingsParams = {
/**
 * @nullable
 */
tariffZone?: string | null;
/**
 * @nullable
 */
bookingType?: string | null;
/**
 * @nullable
 */
status?: string | null;
};