# lib/api-zod/src/generated/types/adminBookingDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */
import type { Sibling } from './sibling';

export interface AdminBookingDetail {
  id: number;
  referenceNumber: string;
  childName: string;
  childAddress: string;
  /** @nullable */
  studentNumber?: string | null;
  gradeYear: string;
  parentName: string;
  parentEmail: string;
  parentPhone: string;
  tariffZone: string;
  bookingType: string;
  outboundRoute: string;
  returnRoute: string;
  signatureName?: string;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
  siblings: Sibling[];
}