# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */
import type { AdminBooking } from './adminBooking';
import type { AdminStatsByBookingType } from './adminStatsByBookingType';
import type { AdminStatsByGradeYear } from './adminStatsByGradeYear';
import type { AdminStatsByStatus } from './adminStatsByStatus';
import type { AdminStatsByTariffZone } from './adminStatsByTariffZone';

export interface AdminStats {
  totalBookings: number;
  byStatus: AdminStatsByStatus;
  byTariffZone: AdminStatsByTariffZone;
  byBookingType: AdminStatsByBookingType;
  byGradeYear: AdminStatsByGradeYear;
  totalChildren: number;
  recentBookings: AdminBooking[];
}