# lib/api-zod/src/generated/types/bookingConfirmation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */

export interface BookingConfirmation {
  id: number;
  referenceNumber: string;
  message: string;
}