# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */
import type { BookingType } from './bookingType';
import type { RouteOption } from './routeOption';
import type { SiblingInput } from './siblingInput';
import type { TariffZone } from './tariffZone';

export interface BookingInput {
  /** @minLength 2 */
  childName: string;
  /** @minLength 5 */
  childAddress: string;
  /** @nullable */
  studentNumber?: string | null;
  gradeYear: string;
  /** @minLength 2 */
  parentName: string;
  parentEmail: string;
  /** @minLength 6 */
  parentPhone: string;
  tariffZone: TariffZone;
  bookingType: BookingType;
  outboundRoute: RouteOption;
  returnRoute: RouteOption;
  /** @maxItems 3 */
  siblings: SiblingInput[];
  confirmationAccepted: boolean;
  /** @minLength 2 */
  signatureName: string;
  gdprConsent?: boolean;
}