# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  received: 'received',
  reviewed: 'reviewed',
  confirmed: 'confirmed',
  query_open: 'query_open',
} as const;