# lib/api-zod/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface BookingStatusUpdate {
  status?: BookingStatus;
  /** @nullable */
  adminNotes?: string | null;
}