# lib/api-zod/src/generated/types/bookingType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */

export type BookingType = typeof BookingType[keyof typeof BookingType];


export const BookingType = {
  full_year: 'full_year',
  first_half: 'first_half',
} as const;