# lib/api-zod/src/generated/types/routeOption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */

export type RouteOption = typeof RouteOption[keyof typeof RouteOption];


export const RouteOption = {
  zone1: 'zone1',
  zone2: 'zone2',
  zone3: 'zone3',
  none: 'none',
} as const;