# lib/api-zod/src/generated/types/sibling.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */

export interface Sibling {
  id: number;
  childName: string;
  /** @nullable */
  studentNumber?: string | null;
  gradeYear: string;
  outboundRoute: string;
  returnRoute: string;
}