# lib/api-zod/src/generated/types/tariffZone.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regionalshuttle Buchungssystem - Stiftung Louisenlund
 * OpenAPI spec version: 0.1.0
 */

export type TariffZone = typeof TariffZone[keyof typeof TariffZone];


export const TariffZone = {
  zone1: 'zone1',
  zone2: 'zone2',
  zone3: 'zone3',
} as const;